requirements.txt
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
findspark==1.3.0
-e git://github.com/h1-the-swan/h1theswan_utils.git@e20bb37c92997d5f78f20475f2edd9e3b9f7b6f8#egg=h1theswan_utils
humanfriendly==4.17
idna==2.8
infomap==1.0.0b47
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.15.4
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
py4j==0.10.7
pyarrow==0.11.1
Pygments==2.3.1
pyspark==2.4.0
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
